% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlintervals.R
\name{plot.interv_fit}
\alias{plot.interv_fit}
\title{Plotting a interv_fit object}
\usage{
\method{plot}{interv_fit}(x, ..., from = NULL, to = NULL,
  points = 300, plot_bands = TRUE, sigma2 = NULL, col_means = 1,
  col_line = 4, col_bands = rgb(0, 0, 1, 0.25))
}
\arguments{
\item{x}{an object of class interv_fit}

\item{...}{any graphical parameter to be used for plotting}

\item{from, to}{limits of the plotting window of time. By default, it plots the window
corresponding to first 20 time intervals, or the whole series if it is shorter. A warning
will be raised if a part of the series is omitted. By setting these parameters, it is
possible to plot a window of any lenght, but windows too wide will make the graph too
dense}

\item{points}{integer, the number of points in which to evaluate the curve}

\item{plot_bands}{logical; if TRUE, confidence bands are plotted. Default is TRUE}

\item{sigma2}{it is possible to fix scale parameter for confidence bands. If
\code{sigma2} is fixed, then normal quantiles are used; if is not, t quantiles are used}

\item{col_means}{color for interval observed integrals}

\item{col_line}{color for interpolating curve}

\item{col_bands}{color for the region inside confidence bands. Default uses transparency}
}
\description{
Plots the interpolating curve along the interval data. By default, also plots confidence
bands
}
\examples{
M= c(0, 1, -1, 2, 4)
mod= interv_fit(M)
plot(mod)
}
