% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlintervals.R
\name{predict.interv_fit}
\alias{predict.interv_fit}
\title{Extracting point estimates of the interpolated process}
\usage{
\method{predict}{interv_fit}(object, t, se.fit = FALSE, ...)
}
\arguments{
\item{object}{an object of class \code{interv_fit}}

\item{t}{a numeric vector; values must lay between 0 and N, where N corresponds to the
lenght of the data vector on which the model was estimated}

\item{se.fit}{boolean value; if TRUE standard errors of the estimates are returned along
the point estimates}

\item{...}{further arguments that will not be processed}
}
\value{
Object returned depends on the parameter \code{se.fit}. If set to FALSE, a
numerical vector is returned, elsewise a data.frame is returned with a column of estimates
and a column with their SEs
}
\description{
For new values of t provided whitin the scope of the model, the corresponding Maximum
LIkelihood estimated value of the process is returned
}
\examples{
M= c(0, 1, -1, 2, 4)
mod= interv_fit(M, t_0= 0, t_end= 1)
predict(mod, c(0, .25, .5, .75, 1))
}
